+++ School Manger/Models/Bill.cs
@@ -2,8 +2,8 @@
 {
     public class Bill
     {
-        public int Id { get; set; }
-        public int ContractId { get; set; }
+        public long Id { get; set; }
+        public long ContractId { get; set; }
         public long TotalPrice { get; set; }
         public long PaidPrice { get; set; }
         public bool HasPaId {  get; set; }
